+++ src/main/java/lotto/util/NumbersGenerator.java
@@ -1,8 +1,9 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import lotto.domain.Number;
 
 public class NumbersGenerator {
