+++ src/test/java/lotto/util/NumbersGeneratorTest.java
@@ -1,10 +1,11 @@
-package lotto.domain;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import lotto.domain.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -17,7 +18,7 @@ public class NumbersGeneratorTest {
 
         // given
         String[] input = "1, 2, 3, 4, 5, 6".split(", ");
-        List<Number> expected = Arrays.asList(1,2,3,4,5,6)
+        List<Number> expected = Arrays.asList(1, 2, 3, 4, 5, 6)
             .stream()
             .map(Number::new)
             .collect(Collectors.toList());
